Adjusting slope works for yellowlane video
Figured out that slope values have to be limited within a range to
eliminate noise lines. There should be a better way to derive the slope
values, which is hard coded for now.

For the challenge video, image size is different and hence the vertices
of polygon also need to be derived from image size than being hard coded
as of now.

ToDo:
Yet to make all parameters derived than it being hardcoded as of now

There are still some noise lines which can be made robust by deriving
slope values.