Missing lines in some frames of videos fixed
Some of the frames in video had missing line, which is fixed now
varying Hough parameters.

Some of the images did not have very central line which is fixed.

Line detected in one of the frames in the yellow lane video with curved
lane is not very central which is suspected due to
the polyfit straight line curvefitting algorithm used, which is added as
a limitation in the pipeline.

NOTE for REVIEW:

My initial commits for images and videos as they worked had hard coded
values and they worked perfect. It was when I tweaked
parameters for challenge video that best results for initial images and
videos got ruined a bit. I guess I have learned how certain parameters vary differently for different sets of images/frames through the process. I have mentioned
them in the limitations and suggested improvements. Kindly check my
previous commits as well to explain how I reached here.